requirements.txt
geopandas==1.0.1
pandas==2.2.3
pandasai==2.4.2
streamlit==1.44.0
openai==1.69.0
python-dotenv==1.1.0
lida==0.0.14
